Implements 'TOP_6_DIRECTORS_WITHIN_FAMILY' query and fixes a bug in 'GET_TOP_N_MOVIES_RATIO' query

+++ src/pt/ulusofona/deisi/aed/deisiflix/Reader.java
@@ -136,7 +136,7 @@ public class Reader {
 
             if (components.length == 3) {
                 int id = Integer.parseInt(components[0].strip());
-                float votesAverage = Float.parseFloat(components[1].strip());
+                double votesAverage = Double.parseDouble(components[1].strip());
                 int votesTotal = Integer.parseInt(components[2].strip());
 
                 if (DEBUG) {
